backend/requirements.txt
django
djangorestframework
dj-rest-auth
django-allauth